add movement with obstacle avoidance - fix measure_left() - fix measure_right() - add is_there_obstacle() - add is_robot_in_target() - add move_with_obstacle_avoidance_do()

+++ math.h
@@ -14,8 +14,9 @@
 uint32_t power(uint16_t, uint8_t);
 double sqrt(uint16_t);
 double atan(float);
-float sin(int a);
-float cos(int a);
+float sin(int16_t a);
+float cos(int16_t a);
+//uint16_t abs(int16_t num);
 
 #endif	/* MATH_H */
